Redo the error stringification to use the error tables.
Mostly what this means that NSS errors (libssl and libnssutil) now have
reasonable description()s, because those strings are static, and the NSPR
errors are no longer hard-coded into this crate.  Also the conversion to
symbolic name, for Debug, uses the tables instead of baking in an extra
copy of the strings.